refactor(progressView): split formatters.js into modular structure
Split the 1,644-line formatters.js into 12 focused modules for better
separation of concerns and maintainability:

**New module structure:**
- `constants.js` - Shared constants and format options
- `timestampUtils.js` - Timestamp formatting and extraction
- `normalizers.js` - Data normalization with documented contracts
- `htmlBuilders.js` - HTML generation utilities
- `markdownRenderer.js` - Markdown processing with KaTeX
- `baseLogFormatter.js` - Banner entry creation and state management
- `taskGroupFormatter.js` - Task group header formatting
- `taskGroupLevel.js` - Hierarchy level configuration (pure data)
- `logFormatters/bannerFormatters.js` - Thinking/scratchpad formatters
- `logFormatters/toolFormatters.js` - Tool use and web search formatters
- `logFormatters/dataFormatters.js` - File list, latexdiff, statistics
- `logFormatters/messageFormatters.js` - User messages and log lines
- `index.js` - Main entry point with LogEntryFormatter class

**Key improvements:**
- Eliminated circular imports by separating TaskGroupLevel
- Consolidated duplicate toggle icon logic via initToggleIcon helper
- Added XSS protection for user-controlled values (varName, sourceDisplay)
- Extracted createToolElement to reduce duplication in tool formatters
- Removed unused legacy methods and dead code
- Consistent null/undefined handling documented in normalizers
