+++ src/progressView/modules/messageHandlers.js
@@ -8,7 +8,7 @@ import {
 import { progressViewDomHandler } from './domHandlers.js';
 import { createThemeHandlers } from './handlers/themeHandlers.js';
 import { appendFormatted } from './utils.js';
-import { getSharedLogEntryFormatter } from './formatters.js';
+import { getSharedLogEntryFormatter } from './formatters/index.js';
 // Local imports - log state
 import { progressViewState } from './progressViewState.js';
 import { BaseWebviewMessageHandler } from '@common/BaseWebviewMessageHandler.js';
